+++ tests/_internal/clients/test_users.py
@@ -36,8 +36,14 @@ class TestDeviceClient(object):
     def test_unicode_username_get_user_by_username_calls_get_with_username(
         self, session, v3_required_session
     ):
-        username = u"我能吞"
+        username = u"您已经发现了秘密信息"
         client = UserClient(session, v3_required_session)
         client.get_user_by_username(username)
         expected_params = {"username": username}
         session.get.assert_called_once_with(USER_URI, params=expected_params)
+
+    def test_get_user_by_id_calls_get_with_uri_and_params(self, session, v3_required_session):
+        client = UserClient(session, v3_required_session)
+        client.get_user_by_id("USER_ID")
+        uri = "{0}/{1}".format(USER_URI, "USER_ID")
+        session.get.assert_called_once_with(uri)